Make hunger restored and health restored overlays more separate
Fixes a few accidental couplings of config settings (e.g. SHOW_FOOD_VALUES_OVERLAY controlling whether or not health was being rendered and whatnot). This is the Fabric version of 12155b784c5961008b9aa0cb7ac600e898110415 in the 1.16-forge branch, but without separating hunger and health rendering into separate passes. The separate passes are not as necessary in Fabric since we can't really properly account for how other mods might affect the rendering of health.