Fix and massively improve worst-case performance of getEstimatedHealthIncrement
Fixes an infinite loop when foodLevel is full and saturationLevel is small enough that, when it is used to increment exhaustionLevel, the representation of exhaustionLevel would not change.

Additionally, instead of looping potentially millions of times just to get exhaustionLevel above the max when foodLevel is full and saturationLevel is small, we calculate the number of iterations that would be necessary and add all the health/exhaustion in one go. This takes the worst-case number of iterations from the millions down to about 18 (these numbers come from some very basic fuzz testing).

Fixes #181